requirements.txt
# DSPy Ebook - Python Dependencies
# For running code examples and exercises

# Core DSPy framework
dspy-ai==2.5.43

# Language Model Providers
openai>=1.0.0
anthropic>=0.8.0

# Environment and Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: For specific examples
# Uncomment as needed for specific chapters/examples

# Vector databases (for RAG examples)
# chromadb>=0.4.0
# weaviate-client>=3.0.0
# pinecone-client>=2.0.0

# Data processing
# pandas>=2.0.0
# numpy>=1.24.0

# Web scraping (for data collection examples)
# beautifulsoup4>=4.12.0
# requests>=2.31.0

# Async support
# aiohttp>=3.9.0

# Monitoring and logging
# mlflow>=2.9.0

# Deployment
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Jupyter Book for documentation
jupyter-book<2.0.0
jupytext>=1.15.0
sphinx>=5.0.0
myst-nb>=0.17.0
